Initial commit of complex Entity instances.
The proposed commit changes the manner in which Entity instances are
handled. The idea is hide the internal mangling of data structures from
the user so that consistency can be maintained.

Large portions of the new API are mainly immutable, the additional
computation cost will be worth the effort, but the low-level types are
now primitive arrays which increase performance to such a degree that
the nett effect is that the new API is still more performant.

Signed-off-by: Gary Pampara <gpampara@gmail.com>